Rename package - splittable totals shipment connector to splittable quote shipment connector - remove from dandelion.json - ...
