+++ bundles/splittable-quote-shipment-connector/codeception.yml
@@ -1,0 +1,22 @@
+namespace: fond-of-oryx/splittable-quote-shipment-connector
+
+suites:
+  unit:
+    path: .
+
+settings:
+  shuffle: true
+  lint: true
+
+bootstrap: _bootstrap.php
+
+paths:
+  tests: tests
+  output: tests/_output
+  support: tests/_support
+  data: tests/_data
+
+coverage:
+  enabled: true
+  include:
+    - src/*.php
