+++ bundles/splittable-quote-shipment-connector/composer.json
@@ -1,0 +1,33 @@
+{
+  "name": "fond-of-oryx/splittable-quote-shipment-connector",
+  "description": "Splittable Quote Shipment Connector Module",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Daniel Rose",
+      "email": "daniel-rose@gmx.de"
+    }
+  ],
+  "require": {
+    "php": ">=7.3",
+    "fond-of-oryx/splittable-quote-extension": "^1.0.0",
+    "spryker/shipment": "^6.2.0 || ^7.0.0 || ^8.0.0"
+  },
+  "require-dev": {
+    "fond-of-codeception/spryker": "^1.0",
+    "spryker/code-sniffer": "^0.12"
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
+  "autoload": {
+    "psr-4": {
+      "FondOfOryx\\": "src/FondOfOryx/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "FondOfOryx\\": "tests/FondOfOryx/",
+      "Generated\\": "src/Generated/"
+    }
+  }
+}
