+++ bundles/splittable-quote-shipment-connector/src/FondOfOryx/Zed/SplittableQuoteShipmentConnector/Business/Expander/SplittedQuoteExpanderInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace FondOfOryx\Zed\SplittableQuoteShipmentConnector\Business\Expander;
+
+use Generated\Shared\Transfer\QuoteTransfer;
+
+interface SplittedQuoteExpanderInterface
+{
+    /**
+     * @param \Generated\Shared\Transfer\QuoteTransfer $splittedQuoteTransfer
+     *
+     * @return \Generated\Shared\Transfer\QuoteTransfer
+     */
+    public function expand(QuoteTransfer $splittedQuoteTransfer): QuoteTransfer;
+}
