+++ bundles/splittable-quote-shipment-connector/src/FondOfOryx/Zed/SplittableQuoteShipmentConnector/Dependency/Facade/SplittableQuoteShipmentConnectorToShipmentFacadeBridge.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace FondOfOryx\Zed\SplittableQuoteShipmentConnector\Dependency\Facade;
+
+use Generated\Shared\Transfer\QuoteTransfer;
+use Generated\Shared\Transfer\ShipmentMethodTransfer;
+use Spryker\Zed\Shipment\Business\ShipmentFacadeInterface;
+
+class SplittableQuoteShipmentConnectorToShipmentFacadeBridge implements
+    SplittableQuoteShipmentConnectorToShipmentFacadeInterface
+{
+    /**
+     * @var \Spryker\Zed\Shipment\Business\ShipmentFacadeInterface
+     */
+    protected $shipmentFacade;
+
+    /**
+     * @param \Spryker\Zed\Shipment\Business\ShipmentFacadeInterface $shipmentFacade
+     */
+    public function __construct(ShipmentFacadeInterface $shipmentFacade)
+    {
+        $this->shipmentFacade = $shipmentFacade;
+    }
+
+    /**
+     * @param int $idShipmentMethod
+     * @param \Generated\Shared\Transfer\QuoteTransfer $quoteTransfer
+     *
+     * @return \Generated\Shared\Transfer\ShipmentMethodTransfer|null
+     */
+    public function findAvailableMethodById(
+        int $idShipmentMethod,
+        QuoteTransfer $quoteTransfer
+    ): ?ShipmentMethodTransfer {
+        return $this->shipmentFacade->findAvailableMethodById($idShipmentMethod, $quoteTransfer);
+    }
+}
