--- bundles/splittable-totals-shipment-connector/codeception.yml
@@ -1,22 +1,0 @@
-namespace: fond-of-oryx/splittable-totals-shipment-connector
-
-suites:
-  unit:
-    path: .
-
-settings:
-  shuffle: true
-  lint: true
-
-bootstrap: _bootstrap.php
-
-paths:
-  tests: tests
-  output: tests/_output
-  support: tests/_support
-  data: tests/_data
-
-coverage:
-  enabled: true
-  include:
-    - src/*.php
