--- bundles/splittable-totals-shipment-connector/src/FondOfOryx/Zed/SplittableTotalsShipmentConnector/Business/Expander/SplittedQuoteExpanderInterface.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace FondOfOryx\Zed\SplittableTotalsShipmentConnector\Business\Expander;
-
-use Generated\Shared\Transfer\QuoteTransfer;
-
-interface SplittedQuoteExpanderInterface
-{
-    /**
-     * @param \Generated\Shared\Transfer\QuoteTransfer $splittedQuoteTransfer
-     *
-     * @return \Generated\Shared\Transfer\QuoteTransfer
-     */
-    public function expand(QuoteTransfer $splittedQuoteTransfer): QuoteTransfer;
-}
