--- bundles/splittable-totals-shipment-connector/src/FondOfOryx/Zed/SplittableTotalsShipmentConnector/Dependency/Facade/SplittableTotalsShipmentConnectorToShipmentFacadeBridge.php
@@ -1,37 +1,0 @@
-<?php
-
-namespace FondOfOryx\Zed\SplittableTotalsShipmentConnector\Dependency\Facade;
-
-use Generated\Shared\Transfer\QuoteTransfer;
-use Generated\Shared\Transfer\ShipmentMethodTransfer;
-use Spryker\Zed\Shipment\Business\ShipmentFacadeInterface;
-
-class SplittableTotalsShipmentConnectorToShipmentFacadeBridge implements
-    SplittableTotalsShipmentConnectorToShipmentFacadeInterface
-{
-    /**
-     * @var \Spryker\Zed\Shipment\Business\ShipmentFacadeInterface
-     */
-    protected $shipmentFacade;
-
-    /**
-     * @param \Spryker\Zed\Shipment\Business\ShipmentFacadeInterface $shipmentFacade
-     */
-    public function __construct(ShipmentFacadeInterface $shipmentFacade)
-    {
-        $this->shipmentFacade = $shipmentFacade;
-    }
-
-    /**
-     * @param int $idShipmentMethod
-     * @param \Generated\Shared\Transfer\QuoteTransfer $quoteTransfer
-     *
-     * @return \Generated\Shared\Transfer\ShipmentMethodTransfer|null
-     */
-    public function findAvailableMethodById(
-        int $idShipmentMethod,
-        QuoteTransfer $quoteTransfer
-    ): ?ShipmentMethodTransfer {
-        return $this->shipmentFacade->findAvailableMethodById($idShipmentMethod, $quoteTransfer);
-    }
-}
