--- bundles/splittable-totals-shipment-connector/tests/FondOfOryx/Zed/SplittableTotalsShipmentConnector/Business/SplittableTotalsShipmentConnectorFacadeTest.php
@@ -1,73 +1,0 @@
-<?php
-
-namespace FondOfOryx\Zed\SplittableTotalsShipmentConnector\Business;
-
-use Codeception\Test\Unit;
-use FondOfOryx\Zed\SplittableTotalsShipmentConnector\Business\Expander\SplittedQuoteExpanderInterface;
-use Generated\Shared\Transfer\QuoteTransfer;
-
-class SplittableTotalsShipmentConnectorFacadeTest extends Unit
-{
-    /**
-     * @var \FondOfOryx\Zed\SplittableTotalsShipmentConnector\Business\SplittableTotalsShipmentConnectorBusinessFactory|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected $businessFactoryMock;
-
-    /**
-     * @var \FondOfOryx\Zed\SplittableTotalsShipmentConnector\Business\Expander\SplittedQuoteExpanderInterface|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected $splittedQuoteExpanderMock;
-
-    /**
-     * @var \Generated\Shared\Transfer\QuoteTransfer|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected $quoteTransferMock;
-
-    /**
-     * @var \FondOfOryx\Zed\SplittableTotalsShipmentConnector\Business\SplittableTotalsShipmentConnectorFacade
-     */
-    protected $facade;
-
-    /**
-     * @return void
-     */
-    protected function _before(): void
-    {
-        parent::_before();
-
-        $this->businessFactoryMock = $this->getMockBuilder(SplittableTotalsShipmentConnectorBusinessFactory::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->splittedQuoteExpanderMock = $this->getMockBuilder(SplittedQuoteExpanderInterface::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->quoteTransferMock = $this->getMockBuilder(QuoteTransfer::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->facade = new SplittableTotalsShipmentConnectorFacade();
-        $this->facade->setFactory($this->businessFactoryMock);
-    }
-
-    /**
-     * @return void
-     */
-    public function testSplittedExpandQuote(): void
-    {
-        $this->businessFactoryMock->expects(static::atLeastOnce())
-            ->method('createSplittedQuoteExpander')
-            ->willReturn($this->splittedQuoteExpanderMock);
-
-        $this->splittedQuoteExpanderMock->expects(static::atLeastOnce())
-            ->method('expand')
-            ->with($this->quoteTransferMock)
-            ->willReturn($this->quoteTransferMock);
-
-        static::assertEquals(
-            $this->quoteTransferMock,
-            $this->facade->expandSplittedQuote($this->quoteTransferMock)
-        );
-    }
-}
