--- bundles/splittable-totals-shipment-connector/tests/FondOfOryx/Zed/SplittableTotalsShipmentConnector/SplittableTotalsShipmentConnectorDependencyProviderTest.php
@@ -1,94 +1,0 @@
-<?php
-
-namespace FondOfOryx\Zed\SplittableTotalsShipmentConnector;
-
-use Codeception\Test\Unit;
-use FondOfOryx\Zed\SplittableTotalsShipmentConnector\Dependency\Facade\SplittableTotalsShipmentConnectorToShipmentFacadeBridge;
-use Spryker\Shared\Kernel\BundleProxy;
-use Spryker\Zed\Kernel\Container;
-use Spryker\Zed\Kernel\Locator;
-use Spryker\Zed\Shipment\Business\ShipmentFacadeInterface;
-
-class SplittableTotalsShipmentConnectorDependencyProviderTest extends Unit
-{
-    /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|\Spryker\Zed\Kernel\Container
-     */
-    protected $containerMock;
-
-    /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|\Spryker\Zed\Kernel\Locator
-     */
-    protected $locatorMock;
-
-    /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|\Spryker\Shared\Kernel\BundleProxy
-     */
-    protected $bundleProxyMock;
-
-    /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|\Spryker\Zed\Shipment\Business\ShipmentFacadeInterface
-     */
-    protected $shipmentFacadeMock;
-
-    /**
-     * @var \FondOfOryx\Zed\SplittableTotalsShipmentConnector\SplittableTotalsShipmentConnectorDependencyProvider
-     */
-    protected $dependencyProvider;
-
-    /**
-     * @return void
-     */
-    protected function _before(): void
-    {
-        parent::_before();
-
-        $this->containerMock = $this->getMockBuilder(Container::class)
-            ->setMethodsExcept(['factory', 'set', 'offsetSet', 'get', 'offsetGet', 'has', 'offsetExists'])
-            ->getMock();
-
-        $this->locatorMock = $this->getMockBuilder(Locator::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->bundleProxyMock = $this->getMockBuilder(BundleProxy::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->shipmentFacadeMock = $this->getMockBuilder(ShipmentFacadeInterface::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->dependencyProvider = new SplittableTotalsShipmentConnectorDependencyProvider();
-    }
-
-    /**
-     * @return void
-     */
-    public function testProvideBusinessLayerDependencies(): void
-    {
-        $this->containerMock->expects(static::atLeastOnce())
-            ->method('getLocator')
-            ->willReturn($this->locatorMock);
-
-        $this->locatorMock->expects(static::atLeastOnce())
-            ->method('__call')
-            ->with('shipment')
-            ->willReturn($this->bundleProxyMock);
-
-        $this->bundleProxyMock->expects(static::atLeastOnce())
-            ->method('__call')
-            ->with('facade')
-            ->willReturn($this->shipmentFacadeMock);
-
-        $container = $this->dependencyProvider->provideBusinessLayerDependencies(
-            $this->containerMock
-        );
-
-        static::assertEquals($this->containerMock, $container);
-        static::assertInstanceOf(
-            SplittableTotalsShipmentConnectorToShipmentFacadeBridge::class,
-            $container[SplittableTotalsShipmentConnectorDependencyProvider::FACADE_SHIPMENT]
-        );
-    }
-}
