+++ composer.json
@@ -126,13 +126,13 @@
     "fond-of-oryx/splittable-checkout-rest-api-shipment-connector": "*",
     "fond-of-oryx/splittable-quote": "*",
     "fond-of-oryx/splittable-quote-extension": "*",
+    "fond-of-oryx/splittable-quote-shipment-connector": "*",
     "fond-of-oryx/splittable-totals": "*",
     "fond-of-oryx/splittable-totals-extension": "*",
     "fond-of-oryx/splittable-totals-rest-api": "*",
     "fond-of-oryx/splittable-totals-rest-api-company-unit-address-connector": "*",
     "fond-of-oryx/splittable-totals-rest-api-extension": "*",
     "fond-of-oryx/splittable-totals-rest-api-shipment-connector": "*",
-    "fond-of-oryx/splittable-totals-shipment-connector": "*",
     "fond-of-oryx/tax-calculation-connector": "*",
     "fond-of-oryx/util-math-formula": "*"
   },
@@ -245,13 +245,13 @@
       "FondOfOryx\\Zed\\SplittableCheckoutRestApiShipmentConnector\\": "bundles/splittable-checkout-rest-api-shipment-connector/src/FondOfOryx/Zed/SplittableCheckoutRestApiShipmentConnector",
       "FondOfOryx\\Zed\\SplittableQuote\\": "bundles/splittable-quote/src/FondOfOryx/Zed/SplittableQuote",
       "FondOfOryx\\Zed\\SplittableQuoteExtension\\": "bundles/splittable-quote-extension/src/FondOfOryx/Zed/SplittableQuoteExtension",
+      "FondOfOryx\\Zed\\SplittableQuoteShipmentConnector\\": "bundles/splittable-quote-shipment-connector/src/FondOfOryx/Zed/SplittableQuoteShipmentConnector",
       "FondOfOryx\\Zed\\SplittableTotalsRestApi\\": "bundles/splittable-totals-rest-api/src/FondOfOryx/Zed/SplittableTotalsRestApi",
       "FondOfOryx\\Zed\\SplittableTotalsRestApiCompanyUnitAddressConnector\\": "bundles/splittable-totals-rest-api-company-unit-address-connector/src/FondOfOryx/Zed/SplittableTotalsRestApiCompanyUnitAddressConnector",
       "FondOfOryx\\Zed\\SplittableTotalsRestApiExtension\\": "bundles/splittable-totals-rest-api-extension/src/FondOfOryx/Zed/SplittableTotalsRestApiExtension",
       "FondOfOryx\\Zed\\SplittableTotalsRestApiShipmentConnector\\": "bundles/splittable-totals-rest-api-shipment-connector/src/FondOfOryx/Zed/SplittableTotalsRestApiShipmentConnector",
       "FondOfOryx\\Zed\\SplittableTotals\\": "bundles/splittable-totals/src/FondOfOryx/Zed/SplittableTotals",
       "FondOfOryx\\Zed\\SplittableTotalsExtension\\": "bundles/splittable-totals-extension/src/FondOfOryx/Zed/SplittableTotalsExtension",
-      "FondOfOryx\\Zed\\SplittableTotalsShipmentConnector\\": "bundles/splittable-totals-shipment-connector/src/FondOfOryx/Zed/SplittableTotalsShipmentConnector",
       "FondOfOryx\\Zed\\TaxCalculationConnector\\": "bundles/tax-calculation-connector/src/FondOfOryx/Zed/TaxCalculationConnector"
     }
   },
@@ -357,13 +357,13 @@
       "FondOfOryx\\Zed\\SplittableCheckoutRestApiShipmentConnector\\": "bundles/splittable-checkout-rest-api-shipment-connector/tests/FondOfOryx/Zed/SplittableCheckoutRestApiShipmentConnector",
       "FondOfOryx\\Zed\\SplittableQuote\\": "bundles/splittable-quote/tests/FondOfOryx/Zed/SplittableQuote",
       "FondOfOryx\\Zed\\SplittableQuoteExtension\\": "bundles/splittable-quote-extension/tests/FondOfOryx/Zed/SplittableQuoteExtension",
+      "FondOfOryx\\Zed\\SplittableQuoteShipmentConnector\\": "bundles/splittable-quote-shipment-connector/tests/FondOfOryx/Zed/SplittableQuoteShipmentConnector",
       "FondOfOryx\\Zed\\SplittableTotalsRestApiCompanyUnitAddressConnector\\": "bundles/splittable-totals-rest-api-company-unit-address-connector/tests/FondOfOryx/Zed/SplittableTotalsRestApiCompanyUnitAddressConnector",
       "FondOfOryx\\Zed\\SplittableTotalsRestApiExtension\\": "bundles/splittable-totals-rest-api-extension/tests/FondOfOryx/Zed/SplittableTotalsRestApiExtension",
       "FondOfOryx\\Zed\\SplittableTotalsRestApiShipmentConnector\\": "bundles/splittable-totals-rest-api-shipment-connector/tests/FondOfOryx/Zed/SplittableTotalsRestApiShipmentConnector",
       "FondOfOryx\\Zed\\SplittableTotalsRestApi\\": "bundles/splittable-totals-rest-api/tests/FondOfOryx/Zed/SplittableTotalsRestApi",
       "FondOfOryx\\Zed\\SplittableTotals\\": "bundles/splittable-totals/tests/FondOfOryx/Zed/SplittableTotals",
       "FondOfOryx\\Zed\\SplittableTotalsExtension\\": "bundles/splittable-totals-extension/tests/FondOfOryx/Zed/SplittableTotalsExtension",
-      "FondOfOryx\\Zed\\SplittableTotalsShipmentConnector\\": "bundles/splittable-totals-shipment-connector/tests/FondOfOryx/Zed/SplittableTotalsShipmentConnector",
       "FondOfOryx\\Zed\\TaxCalculationConnector\\": "bundles/tax-calculation-connector/tests/FondOfOryx/Zed/TaxCalculationConnector",
       "Generated\\": "src/Generated/",
       "Orm\\Zed\\": "src/Orm/Zed/"
